feat: configure the app for development
It leverages Go build tags to override the ConfigureApp function, stating the dev services only for that.

When running in production, you don't want to pass the "-tags=dev" flag to Go, but for dev,
it allows you to create this development experience that, combined with air, makes it super smooth building Go apps.